Write a reversible Decimal to Binary program
This assignment combines knowledge of converting a decimal number to binary, vice-versa
with that of creating a menu-driven, user-input system that will dynamically generate
the correct output given valid inputs.

User input validation is somewhat limited, but this program will try to discard as much
garbage data as possible with a default value of 0.