Fixed bug in ordered pairs loop (#5)
* Make a Matrix Program

Given a set of ordered pairs, the program should create a proper
1-0 matrix to indicate the ordered pairs.

Much of the thinking and work in this program iteration is to
_just get it working first_. I tried to use the simplest structures
I could think of to represent ordered pairs without overly relying on
object oriented features like <map> and <pair> and <vector> although all
of those #includes were considered while building the program.

Future versions of this program would include the option for user input
and maybe even a parser sub-function that can go ahead and accept full int,char
inputs.

* Clarify a comment for clearer documentation purposes

* Fixed bug in ordered pairs loop

I incorrectly set the outermost loop to be the size of setA because, at the time,
the number of ordered pairs and number of elements in setA were the same. Obviously,
I overlooked the fact that loop would break once I added one more ordered pair to the
relations array.

I fixed the bug and also updated the comments to document this.