Priority Queue finish and maintenance
There were some changes that had to be made in the comments
both for clarification and also for the purpose of fulfilling the
assignment--> Making sure the O(n) notation was present for each
and every function that appeared in the classes.

I also started on writing the Double Linked List. Circular arrays soon to follow.