Refactoring LinkedList and Stack
I had commited a certain number of errors in writing these classes
without also keeping in mind readability. So in this refactor pass, I've moved all the function implementations
to the bottom of the header files in Linked List and Stacks and using class scope resolution operators
to relate the function implementations back to the class.

While I was at it, I also decided to add another variant on the traverse method for the stack
class. Mostly for my own educational purposes, but it also seemed like a fun way to show how
inter-node traversal worked.